+++ ListaMatrizes2/matriz4.c
@@ -11,34 +11,34 @@ c[i][j] = 0, se a[i][j] da matriz A não existir em B.
 
 int main(void){
 
-    int matrizA[5][5];
+    float matrizA[5][5];
     int matrizB[5][5];
     int matrizC[5][5];
 
-    srand(time(NULL));      // gerar as 2 matrizes5x5 [codigo de leitura das matrizes comentado abaixo :)]
+    srand(time(NULL));      // gerar as 2 matrizes
     
     for (int i = 0; i < 5; i++) {
         for (int j = 0; j < 5; j++){
-            matrizA[i][j] = rand() % 100;
+            matrizA[i][j] = rand() % 100;       // matrizA
         }
     }
 
     for (int i = 0; i < 5; i++) {
         for (int j = 0; j < 5; j++){
-            matrizB[i][j] = rand() % 100;
+            matrizB[i][j] = rand() % 100;       // matrizB
         }
     }
 
-/*    for (int i = 0; i < 5; i++)
+/*    for (int i = 0; i < 5; i++)               // receber valores matrizA
     {
         for (int j = 0; j < 5; j++)
         {
             printf("Informe o valor correspondente a linha [%d] e coluna [%d]: ", i+1, j+1);
-            scanf("%d", &matrizA[i][j]);
+            scanf("%f", &matrizA[i][j]);
         }
     }
 
-    for (int i = 0; i < 5; i++)
+    for (int i = 0; i < 5; i++)                 // receber valores matrizB
     {
         for (int j = 0; j < 5; j++)
         {
